Authorization과 Login 통합, 비밀번호 변경 후 logout
